Add missing cmake include, also missing code
Turns out I accidentally pushed the IPC logic in the GameHandler so might as well push it completely now.

Documentation might follow though it's pretty similar to sending the data via. the websocket.

+++ src/gtasa/effects/custom/player/StruckByTruckEffect.cpp
@@ -59,7 +59,7 @@ public:
         CVector position
             = player->TransformFromObjectSpace (CVector (0.0f, 25.0f, 0.0f));
 
-        float playerFacing = player->m_fHeadingCurrent - M_PI;
+        float playerFacing = (float) (player->m_fHeadingCurrent - M_PI);
 
         if (linerunner)
         {
